+++ src/main/java/commands/TeamStatManager.java
@@ -1,10 +1,7 @@
 package commands;
 
-import java.util.ArrayList;
-
-import driversAdapters.DataContainer;
 import constants.Exceptions;
-import team.Team;
+import sports.Team;
 
 import java.util.*;
