+++ src/main/java/sports/hockey/HockeyTeamStatManager.java
@@ -2,7 +2,6 @@ package sports.hockey;
 
 import driversAdapters.DataContainer;
 import commands.TeamStatManager;
-import team.HockeyTeam;
 
 import java.util.*;
